added authentication so users may only alter their own notes, and added new error messages for trying to access notes of games that have no junction because user does not have them on their list

+++ src/notes/notes-router.js
@@ -1,22 +1,28 @@
 const express = require('express');
 
 const notesService = require('./notes-service');
+const { requireAuth } = require('../middleware/jwt-auth')
 
 const notesRouter = express.Router();
 const jsonBodyParser = express.json();
 
 notesRouter
   .route('/:gameId/user/:userId')
-  .get((req, res, next) => {
+  .get(requireAuth, (req, res, next) => {
     notesService.findJunction(req.app.get('db'), req.params.userId, req.params.gameId)
-      .then(junction => notesService.getNotes(req.app.get('db'), junction.id))
+      .then(junction => {
+        if (!junction) return res.status(400).json({ error: 'Game not on users game list'})
+        return notesService.getNotes(req.app.get('db'), junction.id)
+      })
       .then(notes => res.json(notes))
       .catch(next)
   })
-  .post(jsonBodyParser, (req, res, next) => {
+  .post(requireAuth, jsonBodyParser, (req, res, next) => {
+    if (req.user.id !== Number(req.params.userId)) return res.status(401).json({ error: 'Unauthorized request'})
     const unverifiedNote = req.body
     notesService.findJunction(req.app.get('db'), req.params.userId, req.params.gameId)
       .then(junction => {
+        if (!junction) return res.status(400).json({ error: 'Game not on users game list'})
         unverifiedNote.junction_id = junction.id
         const note = notesService.verifyNote(unverifiedNote)
         if (typeof note !== 'object') {
@@ -28,13 +34,15 @@ notesRouter
       .then(newNote => res.json(newNote))
       .catch(next)
   })
-  .patch(jsonBodyParser, (req, res, next) => {
+  .patch(requireAuth, jsonBodyParser, (req, res, next) => {
+    if (req.user.id !== Number(req.params.userId)) return res.status(401).json({ error: 'Unauthorized request'})
     const unverifiedNote = req.body
     if (!req.body.id) {
       return res.status(400).json({error: "Missing 'id' in request body"})
     }
     notesService.findJunction(req.app.get('db'), req.params.userId, req.params.gameId)
       .then(junction => {
+        if (!junction) res.status(400).json({ error: 'Game not on users game list'})
         unverifiedNote.junction_id = junction.id
         const note = notesService.verifyNote(unverifiedNote)
         if (typeof note !== 'object') {
@@ -51,7 +59,8 @@ notesRouter
       })
       .catch(next)
   })
-  .delete(jsonBodyParser, (req, res, next) => {
+  .delete(requireAuth, jsonBodyParser, (req, res, next) => {
+    if (req.user.id !== Number(req.params.userId)) return res.status(401).json({ error: 'Unauthorized request'})
     const { id } = req.body;
     
     notesService.deleteNote(req.app.get('db'), id)
